kvm: x86, powerpc: do not allow clearing largepages debugfs entry
The largepages debugfs entry is incremented/decremented as shadow
pages are created or destroyed.  Clearing it will result in an
underflow, which is harmless to KVM but ugly (and could be
misinterpreted by tools that use debugfs information), so make
this particular statistic read-only.

Cc: kvm-ppc@vger.kernel.org
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

Gbp-Pq: Topic bugfix/x86/itlb_multihit
Gbp-Pq: Name 0001-kvm-x86-powerpc-do-not-allow-clearing-largepages-deb.patch