hrtimer: Avoid ABI change in 3.16.80
Forwarded: not-needed

Commit 82a084a03d4b "hrtimer: Store cpu-number in struct
hrtimer_cpu_base" added a new member to hrtimer_cpu_base.  This
structure is visible to modules, though they shouldn't touch it.

The structure is only allocated in the core kernel (statically,
percpu) so it is safe to add a new member at the end.  Do that and
hide it from genksyms.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name hrtimer-avoid-abi-change-in-3.16.80.patch