ipvs: Avoid ABI change in 3.16.80
Forwarded: not-needed

Commit 74b5e1077f6d "ipvs: move old_secure_tcp into struct netns_ipvs"
added a new member to struct netns_ipvs.  This structure is visible to
module.

The structure is only allocated, and the new member used, in the ipvs
module.  So it is safe to add the new member at the end.  Do that and
hide it from genksyms.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name ipvs-avoid-abi-change-in-3.16.80.patch