hso: fix bailout in error case of probe
Origin: https://git.kernel.org/linus/5fcfb6d0bfcda17f0d0656e4e5b3710af2bbaae5

The driver tries to reuse code for disconnect in case
of a failed probe.
If resources need to be freed after an error in probe, the
netdev must not be freed because it has never been registered.
Fix it by telling the helper which path we are in.

Signed-off-by: Oliver Neukum <oneukum@suse.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name hso-fix-bailout-in-error-case-of-probe.patch