certs: check-in the default x509 config file
Origin: https://git.kernel.org/linus/f3a2ba44e93e2c192a872f2705fe66dbf39708d6

When x509.genkey is created, it prints a log:

  Generating X.509 key generation config

..., which is not the ordinary Kbuild log style.

Check-in the default config as certs/default_x509.genkey to make it
readable, and copy it to certs/x509.genkey if it is not present.

The log is shown in the Kbuild style.

  COPY    certs/x509.genkey

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>

Gbp-Pq: Topic features/all
Gbp-Pq: Name certs-check-in-the-default-x509-config-file.patch