x86/cpu: Avoid ABI change for GDS mitigations
Forwarded: not-needed

The changes to support control of GDS mitigation added some #includes
to arch/x86/kernel/cpu/common.c.  This changed the symbol version for
current_task which is defined there, not because of a real ABI change
but because it now includes definitions of more of the types referred
to by struct task_struct.  To avoid the symbol version change, make
those #includes conditional on !__GENKSYMS__.


Gbp-Pq: Topic debian/abi
Gbp-Pq: Name x86-cpu-avoid-abi-change-for-gds-mitigations.patch