kbuild, bpf: Fix BTF reproducibility
pahole 1.22 introduced the -j option for parallel processing, and
kbuild uses it, but this makes its output unreproducible.

pahole 1.27 introduced --btf_features=+reproducible_build which makes
the output reproducible even when -j is used.

Assume that if KBUILD_BUILD_TIMESTAMP is set then a reproducible build
is required, and use both options if available or neither if not.

Signed-off-by: Ben Hutchings <benh@debian.org>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name kbuild-bpf-fix-btf-reproducibility.patch