Export symbols needed by binder
Bug-Debian: https://bugs.debian.org/901492

We want to enable use of the Android binder driver to support
Waydroid, but it should not be built-in as that would waste resources
and increase security attack surface on systems that don't need it.

Export the currently un-exported symbols it depends on.


Gbp-Pq: Topic debian
Gbp-Pq: Name export-symbols-needed-by-binder.patch